Change order of Computer Vision ops in EduKit3.
Since the image was cropped, then Canny was applied, horizontal lines
were detected at the top and bottom of the cropping. I have changed the
order so Canny is run, then the image is cropped, eliminating the issue.

Also, I have changed the crop range from 0 - 0.5 to 0.01 - 0.35, as on
the test vehicle this works much better.